This is a temporary release to test the new simplified API
