+++ Makefile
@@ -21,7 +21,6 @@ test_integration: build
 	@echo --- Running integration tests
 
 	docker run --name mysqlswift \
-	-e MYSQL_SERVER=$(MYSQL_SERVER) \
 	-e MYSQL_ROOT_PASSWORD=my-secret-pw \
 	-d \
 	-p 3306:3306 mysql
