+++ Sources/MySQL/MySQLConnectionPoolProtocol.swift
@@ -4,23 +4,7 @@ public protocol MySQLConnectionPoolProtocol {
 
   init(connectionString: MySQLConnectionString,
        poolSize: Int,
-       defaultCharset: String,
-       provider: @escaping () -> MySQLConnectionProtocol?)
-
-  /**
-    setConnectionProvider requires you to provide a closure which returns a new object which implements
-    the MySQLConnectionProtocol.  Everytime a new connection is required this block is called.
-
-    - Parameters:
-      - provider: closure returning an object implementing MySQLConnectionProtocol
-
-    ```
-      MySQLConnectionPoolProtocol.setConnectionProvider() {
-        return MySQLConnection()
-      }
-    ```
-  **/
-  func setConnectionProvider(provider: @escaping () -> MySQLConnectionProtocol?)
+       defaultCharset: String)
 
   /**
     setLogger sets the function that will be called when an event occurs with the connectionPool,
@@ -35,24 +19,27 @@ public protocol MySQLConnectionPoolProtocol {
   func setLogger(logger: @escaping (_ message: MySQLConnectionPoolMessage) -> Void)
 
   /**
-    getConnection returns a connection from the pool, if a connection is unsuccessful then getConnection throws a MySQLError,
-    if the pool has no available connections getConnection will block util either a connection is free or a timeout occurs.
+    getClient returns a client from the pool, if a connection is unsuccessful 
+    then getConnection throws a MySQLError. If the pool has no available connections 
+    getConnection will block util either a connection is free or a timeout occurs.
 
-    - Returns: An object implementing the MySQLConnectionProtocol.
+    - Returns: An object implementing the MySQLClientProtocol.
   */
   func getConnection() throws -> MySQLConnectionProtocol?
 
   /**
-    getConnection returns a connection from the pool, if a connection is unsuccessful then getConnection throws a MySQLError,
-    if the pool has no available connections getConnection will block util either a connection is free or a timeout occurs.
+    getClient returns a client from the pool, if a connection is unsuccessful 
+    then getClient throws a MySQLError. If the pool has no available connections 
+    getConnection will block util either a connection is free or a timeout occurs.
 
-    By passing the optional closure once the code has executed within the block the connection is automatically released
-    back to the pool saving the requirement to manually call releaseConnection.
+    By passing the optional closure once the code has executed within the block 
+    the connection is automatically released back to the pool saving the 
+    requirement to manually call releaseConnection.
 
     - Parameters:
       - closure: Code that will be executed before connection is released back to the pool
 
-    - Returns: An object implementing the MySQLConnectionProtocol.
+    - Returns: An object implementing the MySQLClientProtocol.
 
     ```
       MySQLConnectionPoolProtocol.getConnection() {
@@ -62,13 +49,13 @@ public protocol MySQLConnectionPoolProtocol {
       }
     ```
   */
-  func getConnection(closure: ((_: MySQLConnectionProtocol) -> Void)) throws
+  func getConnection(closure: ((_: MySQLConnectionProtocol) throws -> Void)) throws
 
   /**
-    releaseConnection returns a connection to the pool.
+    releaseClient returns a client to the pool.
 
     - Parameters:
-      - connection: Connection to be returned to the pool
+      - clent: Client to be returned to the pool
   */
   func releaseConnection(_ connection: MySQLConnectionProtocol)
 }
