+++ Sources/MySQL/mocks/MockMySQLInternalConnection.swift
@@ -1,0 +1,125 @@
+import Foundation
+/*
+// MockMySQLConnection is a mock object which can be used in unit tests to replace the real instance in order to test behaviour
+public class MockInternalMySQLConnection : MySQLInternalConnectionProtocol {
+  public var isConnectedReturn = true
+  public var isConnectedCalled = false
+  public var connectError: MySQLError?
+
+  public var connectCalled = false
+  public var executeCalled = false
+  public var closeCalled = false
+
+  public var executeStatement = ""
+
+  public var clientInfo: String? = nil
+  public var clientVersion: UInt = 0
+  public var defaultCharset: String = "utf8"
+
+  public var executeReturnResult: CMySQLResult? = nil
+  public var executeReturnHeaders: [CMySQLField]? = nil
+  public var executeReturnError: MySQLError? = nil
+
+  public var nextResultReturn: CMySQLRow? = nil
+
+  public var nextResultReturnResult: CMySQLResult? = nil
+  public var nextResultReturnHeaders: [CMySQLField]? = nil
+  public var nextResultReturnError: MySQLError? = nil
+
+  private var uuid: Double
+
+  public func equals(otherObject: MySQLConnectionProtocol) -> Bool {
+    return uuid == (otherObject as! MockMySQLConnection).uuid
+  }
+
+  public init() {
+    uuid = NSDate().timeIntervalSince1970
+  }
+
+  public func isConnected() -> Bool {
+    isConnectedCalled = true
+    return isConnectedReturn
+  }
+
+  public func connect(
+    host: String,
+    user: String,
+    password: String
+  ) throws {
+    if connectError != nil {
+      throw connectError!
+    }
+
+    connectCalled = true
+  }
+
+  public func connect(
+    host: String,
+    user: String,
+    password: String,
+    port: Int
+  ) throws {
+    if connectError != nil {
+      throw connectError!
+    }
+
+    connectCalled = true
+  }
+
+  public func connect(
+    host: String,
+    user: String,
+    password: String,
+    port: Int,
+    database: String
+  ) throws {
+    if connectError != nil {
+      throw connectError!
+    }
+
+    connectCalled = true
+  }
+
+  public func connect(
+    host: String,
+    user: String,
+    password: String,
+    port: Int,
+    database: String,
+    charset: String
+  ) throws {
+    if connectError != nil {
+      throw connectError!
+    }
+
+    connectCalled = true
+  }
+
+  public func client_info() -> String? { return clientInfo }
+
+  public func client_version() -> UInt { return clientVersion }
+
+  public func charset() -> String? { return defaultCharset }
+
+  public func setCharset(charset: String) -> Bool {
+     self.defaultCharset = charset
+     return true
+  }
+
+  public func execute(query: String) -> (CMySQLResult?, [CMySQLField]?, MySQLError?) {
+     executeCalled = true
+     executeStatement = query
+     return (executeReturnResult, executeReturnHeaders, executeReturnError)
+   }
+
+   public func nextResult(result: CMySQLResult) -> CMySQLRow? {
+     return nextResultReturn
+   }
+
+   public func nextResultSet() -> (CMySQLResult?, [CMySQLField]?, MySQLError?) {
+     return (nextResultReturnResult, nextResultReturnHeaders, nextResultReturnError)
+   }
+
+  public func close() { closeCalled = true }
+}
+*/
